+++ src/Codec/CBOR/Cuddle/CDDL/CTreePhase.hs
@@ -12,7 +12,7 @@ module Codec.CBOR.Cuddle.CDDL.CTreePhase (
 ) where
 
 import Codec.CBOR.Cuddle.CDDL (XCddl, XRule, XTerm, XXTopLevel, XXType2)
-import Codec.CBOR.Cuddle.CDDL.Custom.Core (WrappedTerm)
+import Codec.CBOR.Cuddle.CDDL.Custom.Core (RuleTerm)
 import Codec.CBOR.Cuddle.CDDL.Custom.Generator (CBORGen)
 import Codec.CBOR.Cuddle.CDDL.Custom.Validator (TermValidator)
 import Data.Default.Class (Default)
@@ -32,7 +32,7 @@ data instance XCddl CTreePhase = CTreeXCddl
   deriving (Generic, Show, Eq, Ord)
 
 data instance XRule CTreePhase
-  = CTreeXRule (Maybe (CBORGen WrappedTerm)) (Maybe TermValidator)
+  = CTreeXRule (Maybe (CBORGen RuleTerm)) (Maybe TermValidator)
   deriving (Generic)
 
 data instance XXType2 CTreePhase