+++ src/Codec/CBOR/Cuddle/CDDL/Custom/Generator.hs
@@ -18,7 +18,7 @@ module Codec.CBOR.Cuddle.CDDL.Custom.Generator (
 
 import Codec.CBOR.Cuddle.CDDL (GRef (..), Name (..))
 import Codec.CBOR.Cuddle.CDDL.CTree (CTree, CTreeRoot (..), XXCTree)
-import Codec.CBOR.Cuddle.CDDL.Custom.Core (MonadCddl (..), WrappedTerm)
+import Codec.CBOR.Cuddle.CDDL.Custom.Core (MonadCddl (..), RuleTerm)
 import Control.Monad.Reader (MonadReader (..), ReaderT (..), asks, mapReaderT)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
@@ -97,7 +97,7 @@ withLocalGenBindings binds =
 
 data instance XXCTree GenPhase
   = GenRef Name
-  | GenGenerator (CBORGen WrappedTerm) (CTree GenPhase)
+  | GenGenerator (CBORGen RuleTerm) (CTree GenPhase)
 
 class HasGenerator a where
-  generatorL :: Lens' a (Maybe (CBORGen WrappedTerm))
+  generatorL :: Lens' a (Maybe (CBORGen RuleTerm))